Fix: copy mistake that conjured up hard to detect errors.
By a copy mistake, we accessed a variable named text, that was undefined. However, MATLAB does provide a function named text which was executed on this occasion. This function involuntarily opens a figure, if none is open. This lets subsequent mlunit gui tests fail for some obscure reason, but only on recent MATLAB releases (R2014b and newer). On older releases, this works just fine, without opening a figure during the test.

Incidentally, calling the text function did not evoke any other errors, even though it is called within an assertion. That is because the text function returns a vector of graphics handles, which, when being called on a new figure, contains exactly one item, thus satisfying the assertion.

We have to conclude that MATLAB wants us on our toes when writing code, never at ease, and always doubt the code to do what it was intended to do.