Fix: Warning was generated since R2016a for every test suite that doesn't define suite_setup or suite_teardown. This is fixed now.
Problem:

If suite_setup or suite_teardown are not defined in a test suite, the suite_setup_obj and suite_teardown_obj are initialized to

>> suite_setup_obj = function_test_case(0,0,0,'');
>> suite_teardown_obj = function_test_case(0,0,0,'');

in build_testsuite_object.m. The first parameter (0) is the test_function handle, which means that test_function handle is 0 if suite_setup or suite_teardown is not defined in test suite.
This test function is called via

>> self.test_function();

in @function_test_case/run_test.m. If test_function is 0 this generates a Warning since R2016a (and will generate an error in following releases):

"Warning: A value of class "double" was indexed with no subscripts specified. Currently the result of this operation is the indexed value itself, but in a future release, it will be an error."

Fix:

Safeguarded call to self.test_function by

if self.test_function ~= 0
    self.test_function();
end