Add mlunit_variation, providing parameter variations for tests.
A test can call mlunit_variation to define a dynamic or static
list of parameter configurations that are used to call the same
function several times, but with different parameters.
If any of the variations fail or raise an error, the others
continue to run. The test result will be the worst result of
its variations.